+++ NEWS.md
@@ -1,7 +1,8 @@
-# colourlookup 0.1.1.9002  2024-06-16
+# colourlookup 0.1.1.9003  2024-06-18
 
 * Lookup packed integer RGBA colours for nativerasters
 * Include NA and 'transparent' in named colour lookup
+* Handle logical NAs as transparent
 
 # colourlookup 0.1.1  2024-06-16